Program and syntax update
Updated program to:
- use the executable's current directory as Papyrus directory, if
nothing is typed by the user
- create a "snippets" subfolder for output in the Papyrus directory, if
nothing is typed by the user
- create a file called FunctionLog.txt, the contents of which can be
copy-pasted into the functions section of the Papyrus syntax definition,
in the output directory
- create a file called ClassLog.txt, the contents of which can be
copy-pasted into the class section of the Papyrus syntax definition, in
the output directory

If one wants to update the syntax definition, then one should place all
Papyrus source files into a single folder and generate the FunctionLog
and ClassLog files.

SKI_QuestBase.OnInit snippet was also removed and replaced with a
generic OnInit snippet as the event is available to all scripts.

The Papyrus syntax definition was updated to include keywords for
funtions and classes from Skyim (1.9.32.0.8), SKSE (1.7.0 alpha) and
SkyUI SDK (4.1).