Tweaked the way imports work
The folder containing the source file, which is to be compiled, is
loaded first, but only if the folder is not the same as the one set to
contain the vanilla Skyrim source files. Then the additional import
folders are loaded before finally loading the folder containing vanilla
Skyrim source files.