Program update
Added support for reading the first line in "Description.txt" files that
exist in a folder containing .psc files. The line is then added to the
description of a function or event inside parentheses.

Also renamed the source file so that it automatically generates an
executable with the correct name when compiling. Just for convenience.