New feature and minor changes
- Removed HasExtension function and replaced it with str.endswith in the
plugin.
- Added the ability to open any Papyrus script via "SublimePapyrus: Open
script" command. Opens an input panel that accepts regular expressions
(Python reg. ex.). The final pattern, if the input is "actor", would
become "^(actor.psc)$". It is not necessary to include "^", "$", nor
".psc" as those are added if missing. If there are multiple matches,
then a list of matches is shown that the user can choose from or ignore.
- Removed an instance of "INI" from the old command.