Added setting
Added the ability to set the path to the INI file in case you want to
keep it somewhere outside of My Documents. The path needs to be absolute
and can either be the path to a folder, in which case a file called
SublimePapyrus.ini will be created when the INI creation command is run,
or to a file, in which case that file will be used. If nothing is
defined by the user, then the package will use the default path.