Updated Skyrim's plugin and linter
Opening files in Python defaults to using the preferred encoding of the
OS' locale, which may cause issues if the file being read contains bytes
that are undefined in said encoding. The changes made in this commit
will result in an attempt to open the file with UTF-8 encoding in the
event that a UnicodeDecodeError is raised in the initial attempt to read
the file.